altblog/__init__.py
from flask import Flask
from altblog.config import Config
from flask_sqlalchemy import SQLAlchemy
from flask_migrate import Migrate

db = SQLAlchemy()
migrate = Migrate()


def create_app(config_class=Config):
    app = Flask(__name__)
    app.config.from_object(Config)

    db.init_app(app)
    migrate.init_app(app, db)

    # register blueprint here
    from altblog.main.routes import main
    app.register_blueprint(main)

    return app
