requirements.txt
autopep8==1.5.5
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycodestyle==2.6.0
python-dotenv==0.15.0
SQLAlchemy==1.3.17
toml==0.10.2
Werkzeug==1.0.1
